edutech_backend/requirements.txt
# FastAPI framework and server
fastapi==0.110.0
uvicorn[standard]==0.27.1
starlette==0.36.3

# Database
sqlalchemy==2.0.27
alembic==1.13.1
aiomysql==0.2.0
pymysql==1.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.7

# Validation and settings
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# File handling
aiofiles==23.2.1
python-magic==0.4.27

# Utilities
python-dotenv==1.0.1
httpx==0.25.2
tenacity==8.2.3

# AI/ML
openai==1.51.0

# Development tools
black==23.11.0
isort==5.12.0
pytest==7.4.3
pytest-asyncio==0.21.1
ruff==0.1.8
